# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateStripeConnectRequest:
    properties:
      buisinessUrl:
        type: string
      dobDay:
        type: integer
      dobMonth:
        type: integer
      dobYear:
        type: integer
      email:
        type: string
      firstName:
        type: string
      firstNameKana:
        type: string
      lastName:
        type: string
      lastNameKana:
        type: string
      line1:
        type: string
      line2:
        type: string
      line2Kana:
        type: string
      mcc:
        type: string
      phone:
        type: string
      postalCode:
        type: string
      productionDescription:
        type: string
    required:
    - buisinessUrl
    - dobDay
    - dobMonth
    - dobYear
    - email
    - firstName
    - firstNameKana
    - lastName
    - lastNameKana
    - line1
    - line2
    - line2Kana
    - mcc
    - phone
    - postalCode
    - productionDescription
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: SHAPS API
  license:
    name: yuta
  title: SHAPS API
  version: version(1.0)
paths:
...
  /me/stripe-connect:
    post:
      consumes:
      - application/json
      description: create stripeconnect
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStripeConnectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Create StripeConnect
      tags:
      - me
...
swagger: "2.0"